+++ src/Pages/Dashboard/MyPurcess/MypurcessCart.js
@@ -1,30 +1,39 @@
-import { Button, Card, CardContent, CardMedia, Grid, IconButton, Typography } from '@mui/material';
-import { Box } from '@mui/system';
-import React from 'react';
+import {
+  Button,
+  Card,
+  CardContent,
+  CardMedia,
+  Grid,
+  IconButton,
+  Typography,
+} from "@mui/material";
+import { Box } from "@mui/system";
+import React from "react";
 
-const MypurcessCart = ({ bike, setBikes ,bikes}) => {
+const MypurcessCart = ({ bike, setBikes, bikes }) => {
   const handelCancelOrder = (id) => {
-    fetch(`http://localhost:5000/orderdBiek/${id}`, {
-      method: "DELETE",
-    })
-      .then((res) => res.json())
-      .then((data) => {
-        if (data.deletedCount) {
+    const confirm = window.confirm(`Are You Sure,  Want To  Cancel This Order`);
+    if (confirm) {
+      fetch(`http://localhost:5000/orderdBiek/${id}`, {
+        method: "DELETE",
+      })
+        .then((res) => res.json())
+        .then((data) => {
+          if (data.deletedCount) {
             alert("Delete successfully");
-            const remaing = bikes?.filter(bk => bk._id !== id);
-            setBikes(remaing)
-        }
-      });
+            const remaing = bikes?.filter((bk) => bk._id !== id);
+            setBikes(remaing);
+          }
+        });
+    }
   };
 
   const {
-    BuyerName,
-    Buyeremail,
+  
     BiekName,
     Price,
     photo,
-    PhoneNumber,
-    address,
+  
     status,
     Engine,
     _id,
@@ -32,7 +41,7 @@ const MypurcessCart = ({ bike, setBikes ,bikes}) => {
   return (
     <Grid item xs={12} md={6}>
       <Card sx={{ display: "flex" }}>
-        <Box sx={{ display: "flex", textAlign: "left",marginLeft:"5px" }}>
+        <Box sx={{ display: "flex", textAlign: "left", marginLeft: "5px" }}>
           <Grid container>
             <Grid item xs={12} sm={5}>
               <img style={{ width: "100%" }} src={photo} alt=" " />
